icons/map - rename MapChart as Map & update design
