--- src/icons/map-chart.js
@@ -1,34 +1,0 @@
-import React from 'react'
-import PropTypes from 'prop-types'
-
-
-const iconSize = Object.freeze({
-  lg: 'w-3.5 h-3.5',
-  md: 'w-3 h-3',
-  sm: 'w-2.5, h-2.5',
-})
-const MapChart = ({ className, size, ...props }) => {
-  return (
-    <svg
-      className={`${iconSize[size]} ${className}`}
-      viewBox="0 0 15 15"
-      xmlns="http://www.w3.org/2000/svg"
-      fill="currentColor"
-      {...props}
-    >
-      <path
-        d="M11.875 6.87512C11.0462 6.87512 10.2513 7.20436 9.66529 7.79041C9.07924 8.37646 8.75 9.17132 8.75 10.0001C8.75 11.9814 11.5375 14.7876 11.6562 14.9064C11.6843 14.9363 11.7182 14.9601 11.7558 14.9764C11.7934 14.9927 11.834 15.0011 11.875 15.0011C11.916 15.0011 11.9566 14.9927 11.9942 14.9764C12.0318 14.9601 12.0657 14.9363 12.0938 14.9064C12.2125 14.7876 15 11.9814 15 10.0001C15 9.58974 14.9192 9.18338 14.7621 8.80424C14.6051 8.42509 14.3749 8.0806 14.0847 7.79041C13.7945 7.50023 13.45 7.27004 13.0709 7.113C12.6917 6.95595 12.2854 6.87512 11.875 6.87512ZM11.875 11.0939C11.6587 11.0939 11.4472 11.0297 11.2673 10.9095C11.0875 10.7894 10.9473 10.6185 10.8645 10.4187C10.7817 10.2188 10.7601 9.99891 10.8023 9.78674C10.8445 9.57458 10.9486 9.37969 11.1016 9.22672C11.2546 9.07376 11.4495 8.96959 11.6616 8.92739C11.8738 8.88519 12.0937 8.90685 12.2936 8.98963C12.4934 9.07241 12.6642 9.2126 12.7844 9.39247C12.9046 9.57233 12.9688 9.7838 12.9688 10.0001C12.9688 10.2902 12.8535 10.5684 12.6484 10.7735C12.4433 10.9786 12.1651 11.0939 11.875 11.0939Z"
-        strokeWidth="0"
-      />
-      <path 
-        d="M8.75 5.62514C8.75 5.7909 8.81585 5.94988 8.93306 6.06709C9.05027 6.1843 9.20924 6.25014 9.375 6.25014C9.54076 6.25014 9.69973 6.1843 9.81694 6.06709C9.93415 5.94988 10 5.7909 10 5.62514V2.06889C10.0005 2.04014 10.009 2.01209 10.0244 1.98783C10.0398 1.96357 10.0617 1.94404 10.0875 1.93139C10.1124 1.91772 10.1403 1.91055 10.1688 1.91055C10.1972 1.91055 10.2251 1.91772 10.25 1.93139L12.9938 3.89389C13.0342 3.92331 13.0672 3.96184 13.09 4.00636C13.1128 4.05087 13.1248 4.10013 13.125 4.15014V5.62514C13.125 5.7909 13.1908 5.94988 13.3081 6.06709C13.4253 6.1843 13.5842 6.25014 13.75 6.25014C13.9158 6.25014 14.0747 6.1843 14.1919 6.06709C14.3092 5.94988 14.375 5.7909 14.375 5.62514V3.83764C14.3748 3.68838 14.339 3.54132 14.2705 3.4087C14.202 3.27607 14.1028 3.16172 13.9813 3.07514L9.91875 0.175145C9.75997 0.0620974 9.56991 0.00134798 9.375 0.00134798C9.18009 0.00134798 8.99003 0.0620974 8.83125 0.175145L5.2 2.76264C5.14731 2.8037 5.08242 2.82599 5.01563 2.82599C4.94883 2.82599 4.88394 2.8037 4.83125 2.76264L1.51875 0.200145C1.37934 0.0918532 1.21251 0.0245485 1.03698 0.00578107C0.861445 -0.0129864 0.684157 0.0175259 0.525 0.093895C0.367675 0.171788 0.235195 0.292019 0.142451 0.441073C0.0497063 0.590126 0.000374366 0.762094 0 0.937645L0 10.2439C0.00048125 10.3894 0.0345785 10.5327 0.0996292 10.6628C0.16468 10.7929 0.258923 10.9062 0.375 10.9939L4.44375 14.0439C4.60603 14.1656 4.8034 14.2314 5.00625 14.2314C5.20097 14.229 5.39046 14.1681 5.55 14.0564L8.28125 12.1001C8.41076 12.0049 8.49861 11.8635 8.52654 11.7052C8.55448 11.5469 8.52035 11.3839 8.43125 11.2501C8.38388 11.1825 8.3236 11.125 8.25387 11.0808C8.18415 11.0367 8.10636 11.0067 8.02501 10.9928C7.94366 10.9789 7.86035 10.9812 7.77989 10.9996C7.69944 11.018 7.62343 11.0522 7.55625 11.1001L5.875 12.2876C5.8501 12.3013 5.82216 12.3085 5.79375 12.3085C5.76534 12.3085 5.7374 12.3013 5.7125 12.2876C5.68807 12.2765 5.66706 12.2591 5.65167 12.2371C5.63628 12.2151 5.62707 12.1894 5.625 12.1626V4.16264C5.62524 4.11263 5.63724 4.06337 5.66005 4.01886C5.68285 3.97434 5.7158 3.93581 5.75625 3.90639L8.5 1.94389C8.5249 1.93022 8.55284 1.92305 8.58125 1.92305C8.60966 1.92305 8.6376 1.93022 8.6625 1.94389C8.68833 1.95654 8.71016 1.97607 8.7256 2.00033C8.74104 2.02459 8.74948 2.05264 8.75 2.08139V5.62514ZM4.375 12.1189C4.37522 12.1478 4.36708 12.1762 4.35155 12.2006C4.33603 12.225 4.31379 12.2443 4.2875 12.2564C4.2626 12.2701 4.23466 12.2772 4.20625 12.2772C4.17784 12.2772 4.1499 12.2701 4.125 12.2564L1.375 10.1939C1.33654 10.1645 1.3053 10.1266 1.28364 10.0833C1.26198 10.04 1.25047 9.99232 1.25 9.94389V1.87514C1.25 1.84613 1.25808 1.81768 1.27334 1.793C1.28859 1.76832 1.31042 1.74837 1.33637 1.73539C1.36233 1.72241 1.39138 1.71692 1.42028 1.71953C1.44918 1.72213 1.47679 1.73273 1.5 1.75014L4.25625 3.87514C4.29606 3.9058 4.32771 3.94578 4.34842 3.99156C4.36913 4.03734 4.37826 4.08751 4.375 4.13764V12.1189Z"
-        strokeWidth="0"
-      />
-    </svg>
-  )
-}
-
-MapChart.propTypes = { className: PropTypes.string, size: PropTypes.string }
-MapChart.defaultProps = { className: '', size: '' }
-
-export default MapChart
