+++ src/icons/map.js
@@ -1,0 +1,33 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+
+
+const iconSize = Object.freeze({
+  lg: 'w-3.5 h-3.5',
+  md: 'w-3 h-3',
+  sm: 'w-2.5, h-2.5',
+})
+
+const Map = ({ className, size, ...props }) => {
+  return (
+    <svg
+      className={`${iconSize[size]} ${className}`}
+      viewBox="0 0 20 20"
+      xmlns="http://www.w3.org/2000/svg"
+      fill="currentColor"
+      {...props}
+    >
+      <path
+        d="M15.8333 9.16666C14.7283 9.16666 13.6685 9.60565 12.8871 10.387C12.1057 11.1684 11.6667 12.2283 11.6667 13.3333C11.6667 15.975 15.3833 19.7167 15.5417 19.875C15.5791 19.9149 15.6243 19.9466 15.6744 19.9683C15.7246 19.9901 15.7787 20.0013 15.8333 20.0013C15.888 20.0013 15.9421 19.9901 15.9922 19.9683C16.0424 19.9466 16.0876 19.9149 16.125 19.875C16.2833 19.7167 20 15.975 20 13.3333C20 12.7862 19.8922 12.2443 19.6828 11.7388C19.4734 11.2333 19.1665 10.774 18.7796 10.387C18.3927 10.0001 17.9334 9.69322 17.4278 9.48383C16.9223 9.27443 16.3805 9.16666 15.8333 9.16666ZM15.8333 14.7917C15.5449 14.7917 15.2629 14.7061 15.0231 14.5459C14.7833 14.3856 14.5964 14.1579 14.486 13.8914C14.3756 13.6249 14.3468 13.3317 14.403 13.0488C14.4593 12.7659 14.5982 12.5061 14.8021 12.3021C15.0061 12.0982 15.2659 11.9593 15.5488 11.903C15.8317 11.8467 16.1249 11.8756 16.3914 11.986C16.6579 12.0964 16.8857 12.2833 17.0459 12.5231C17.2061 12.7629 17.2917 13.0449 17.2917 13.3333C17.2917 13.7201 17.138 14.091 16.8645 14.3645C16.591 14.638 16.2201 14.7917 15.8333 14.7917Z"
+      />
+      <path
+        d="M11.6667 7.49991C11.6667 7.72092 11.7545 7.93288 11.9107 8.08916C12.067 8.24544 12.279 8.33324 12.5 8.33324C12.721 8.33324 12.933 8.24544 13.0893 8.08916C13.2455 7.93288 13.3333 7.72092 13.3333 7.49991V2.75824C13.334 2.71991 13.3453 2.68251 13.3659 2.65016C13.3865 2.61781 13.4156 2.59177 13.45 2.57491C13.4832 2.55667 13.5205 2.54711 13.5583 2.54711C13.5962 2.54711 13.6335 2.55667 13.6667 2.57491L17.325 5.19158C17.3789 5.2308 17.4229 5.28217 17.4533 5.34152C17.4837 5.40088 17.4997 5.46655 17.5 5.53324V7.49991C17.5 7.72092 17.5878 7.93288 17.7441 8.08916C17.9004 8.24544 18.1123 8.33324 18.3333 8.33324C18.5543 8.33324 18.7663 8.24544 18.9226 8.08916C19.0789 7.93288 19.1667 7.72092 19.1667 7.49991V5.11657C19.1664 4.91756 19.1186 4.72148 19.0273 4.54464C18.936 4.36781 18.8038 4.21534 18.6417 4.09991L13.225 0.233242C13.0133 0.0825117 12.7599 0.00151247 12.5 0.00151247C12.2401 0.00151247 11.9867 0.0825117 11.775 0.233242L6.93333 3.68324C6.86308 3.73798 6.77656 3.7677 6.6875 3.7677C6.59844 3.7677 6.51192 3.73798 6.44167 3.68324L2.025 0.266575C1.83912 0.122186 1.61668 0.0324465 1.38263 0.00742326C1.14859 -0.0176 0.91221 0.0230831 0.7 0.124909C0.490234 0.228766 0.313593 0.389074 0.189934 0.587812C0.0662751 0.78655 0.000499154 1.01584 0 1.24991V13.6582C0.000641667 13.8522 0.0461047 14.0434 0.132839 14.2168C0.219573 14.3903 0.345231 14.5414 0.5 14.6582L5.925 18.7249C6.14137 18.8872 6.40454 18.9749 6.675 18.9749C6.93463 18.9718 7.18728 18.8905 7.4 18.7416L11.0417 16.1332C11.2143 16.0063 11.3315 15.8177 11.3687 15.6066C11.406 15.3956 11.3605 15.1783 11.2417 14.9999C11.1785 14.9098 11.0981 14.8331 11.0052 14.7742C10.9122 14.7153 10.8085 14.6754 10.7 14.6568C10.5915 14.6382 10.4805 14.6413 10.3732 14.6658C10.2659 14.6904 10.1646 14.736 10.075 14.7999L7.83333 16.3832C7.80014 16.4015 7.76288 16.411 7.725 16.411C7.68713 16.411 7.64986 16.4015 7.61667 16.3832C7.58409 16.3684 7.55608 16.3452 7.53556 16.3158C7.51504 16.2865 7.50276 16.2523 7.5 16.2166V5.54991C7.50032 5.48322 7.51633 5.41755 7.54673 5.35819C7.57713 5.29884 7.62107 5.24747 7.675 5.20824L11.3333 2.59158C11.3665 2.57334 11.4038 2.56378 11.4417 2.56378C11.4795 2.56378 11.5168 2.57334 11.55 2.59158C11.5844 2.60844 11.6135 2.63448 11.6341 2.66683C11.6547 2.69917 11.666 2.73657 11.6667 2.77491V7.49991ZM5.83333 16.1582C5.83363 16.1968 5.82277 16.2346 5.80207 16.2671C5.78137 16.2997 5.75171 16.3255 5.71667 16.3416C5.68347 16.3598 5.64621 16.3694 5.60833 16.3694C5.57046 16.3694 5.5332 16.3598 5.5 16.3416L1.83333 13.5916C1.78205 13.5523 1.74039 13.5019 1.71151 13.4442C1.68264 13.3864 1.6673 13.3228 1.66667 13.2582V2.49991C1.66667 2.46122 1.67744 2.42329 1.69778 2.39038C1.71812 2.35747 1.74723 2.33087 1.78183 2.31357C1.81644 2.29627 1.85518 2.28894 1.89371 2.29242C1.93224 2.29589 1.96905 2.31003 2 2.33324L5.675 5.16657C5.72808 5.20745 5.77028 5.26076 5.79789 5.32179C5.82551 5.38283 5.83768 5.44972 5.83333 5.51657V16.1582Z"
+      />
+    </svg>
+  )
+}
+
+Map.propTypes = { className: PropTypes.string, size: PropTypes.string }
+Map.defaultProps = { className: '', size: '' }
+
+export default Map
